move MeasurementOverlayStyle  to /overlayStyle
